reservationForm component váz

+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/ReservationTypeRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.ReservationType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationTypeRepository extends JpaRepository <ReservationType, Integer> {
+public interface ReservationTypeRepository extends JpaRepository <ReservationType, Long> {
 }
